courses crud + authorizations

+++ src/Api/Controllers/CourseController.cs
@@ -1,3 +1,4 @@
+using System.Diagnostics.CodeAnalysis;
 using Bll.Interfaces;
 using Bll.Models;
 using Dal.Models;
@@ -20,7 +21,10 @@ public partial class CourseController(
     [HttpGet("UserCourses")]
     public async Task<IActionResult> GetUserCourses()
     {
-        var result = GetRoleFromBearerToken() switch
+        Role? role = GetRoleFromBearerToken();
+        if (role != Role.Instructor && role != Role.Admin && role != Role.Student) return Unauthorized();
+        if (!ModelState.IsValid) return BadRequest(ModelState);
+        IEnumerable<CourseListDto> result = GetRoleFromBearerToken() switch
         {
             Role.Student => await courseService.GetAllCoursesByStudent(GetUserIdFromBearerToken()!.Value),
             Role.Instructor => await courseService.GetAllCoursesByInstructor(GetUserIdFromBearerToken()!.Value),
@@ -39,6 +43,8 @@ public partial class CourseController(
     [HttpPost("Course")]
     public async Task<IActionResult> CreateCourse([FromBody] CreateCourseDto courseDto)
     {
+        Role? role = GetRoleFromBearerToken();
+        if (role != Role.Instructor && role != Role.Admin) return Unauthorized();
         if (!ModelState.IsValid) return BadRequest(ModelState);
         return Ok(await courseService.CreateCourse(courseDto));
     }
@@ -46,6 +52,8 @@ public partial class CourseController(
     [HttpPut("Course")]
     public async Task<IActionResult> UpdateCourse([FromBody] UpdateCourseDto courseDto)
     {
+        Role? role = GetRoleFromBearerToken();
+        if (role != Role.Instructor && role != Role.Admin) return Unauthorized();
         if (!ModelState.IsValid) return BadRequest(ModelState);
         bool ok = await courseService.UpdateCourse(courseDto);
         return ok ? Ok() : BadRequest();
@@ -54,6 +62,9 @@ public partial class CourseController(
     [HttpDelete("Course/{id:int}")]
     public async Task<IActionResult> DeleteCourse(int id)
     {
+        Role? role = GetRoleFromBearerToken();
+        if (role != Role.Instructor && role != Role.Admin) return Unauthorized();
+        if (!ModelState.IsValid) return BadRequest(ModelState);
         bool ok = await courseService.DeleteCourse(id);
         return ok ? Ok() : BadRequest();
     }
@@ -64,6 +75,7 @@ public partial class CourseController(
         {
             string s = HttpContext.Request.Headers.Authorization.ToString();
             if (!s.StartsWith("Bearer ")) return null;
+            if (!tokenService.ValidateToken(s[7..])) return null;
             return s[7..];
         }
         catch (Exception)
